ml/requirements.txt
# Start lean; install prophet later (heavier, requires build toolchain). Pin minor versions for reproducibility.
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0
python-dotenv==1.0.1
requests==2.31.0
joblib==1.3.2
# Uncomment later when ready for advanced seasonality modeling
# prophet==1.1.5
